Implemented Modules for intcode, aoc2019.Day05
The issue with Day 5 is that opcodes must support parameter mode,
which is not possible using the SimpleAdder and SimpleMultiplier
modules availale. ParameterizedAdder, ... modules were thought of
wherein module.function() takes the burden of checking if it should
work with ic.Current().

Instead of that, a Module now consists of a parameterized bool,
which determines if a Module supports parameter mode,
simplifying the loop in Operate()
and easing burden on Module.function().

Modules implemented are:
(note that pseudocode assumes all in position mode):
* Inputter: 3 ARG1; mem[ARG1] = input; pc+=2
* Outputter: 4 ARG1; output=append(output, mem[ARG1]); pc+=2
* OutputAndHalt: similar to Outputter but if mem[ARG1] != 0 halt
* JumpIfTrue: 5 ARG[1..2]; if mem[ARG1]!=0 jump(mem[ARG2])
* JumpIfFalse: 6 ARG[1..2]; if mem[ARG1]==0 jump(mem[ARG2])
* LessThan: 7 ARG[1..3]; mem[ARG3] = mem[ARG1]<mem[ARG2]  ? 1 : 0
* Equal: 8 ARG[1..3]; mem[ARG3] = mem[ARG1]==mem[ARG2]  ? 1 : 0

For ic.Operate(),
if module.parameterized then check only ic.Current() % 100
otherwise check ic.Current(),
thus no longer needing an module.opcode == 0 if block.

The usual, added Day 5 to README progress,
as well as Day 10 to 17 in aoc2019/readme.