Turned IntCode.input to a stack instead  of an int
Day 7 requires multiple inputs, hence the use of a stack
would be better than that of a single int.

InstallAdderMultiplier and InstallJumpers are functions also added
to make it easier to install the jump modules and adder/multiplier.
Note that no InstallIO is installed
since there are several implementations for outputting (opcode 4).