Added int64 support, dynamic memory allocation, relative mode
"BOOST" programs as described in Day09 requires support for integers
as large as 16 digits long. int64 should be sufficient to support such numbers.

Dynamic memory allocation is also implemented
wherein accessing memory locations beyond the length of ic.mem
will cause ic.mem to append zeroes at the end.

Relative mode is a mode (flag 2) wherein the parameter is interpreted
as a position (akin to position mode) but count from a "relative base".
How getFromMemory does this is that it requires a new parameter,
writeEnable, which determines if memory is needed to be written.
It is assumed always that parameters[len(parameters)-1] is
the address to be written.