Y2015D09: Added more tests + greed + concurrency
shortestPathGreedy is a greedy algorithm that takes in all
points on a graph and determines which point yields the
shortest path using greed: pick the shortest path by choosing
the towns to visit which are shortest to the current one.

shortestPathPermutative is the algorithm that was
originally implemented. It's slow. Please don't.

Day09 still uses *Permutative but uses concurrency.
This should make things fast.

const disconnected signals that a path is not connected.